refactor: domain-layer.jsをclean-architecture/domain.jsに移動
restricted-imports/domain-layer.jsはドメイン層のClean Architecture
制約を表すルールなので、より適切なディレクトリ構造に変更。
