+++ host-frontend-root/frontend-src-root/eslint-rules/clean-architecture/domain.js
